By default pkMultipleSelect assumes the first option is a label and doesn't allow you to actually add it to the selection. This is OK for a lot of our sites, but it's wrong for progressive enhancement of a garden variety multiple select control. And therefore wrong for enhancing Doctrine forms.
So pkMultipleSelect now supports a 'choose-one' option. When you 
set this option to custom label text for the 'CHOOSE ONE' option at
the top of the dropdown, pkMultipleSelect adds an extra option just
for that purpose, and the user is allowed to select ALL of the
"real" options in the original multiple select element.

Also, pkMutlipleSelectAll() now supports an options array too.

Example:

pkMultipleSelectAll({'choose-one': 'Select to Add'});




git-svn-id: http://svn.symfony-project.com/plugins/pkToolkitPlugin/trunk@21706 ee427ae8-e902-0410-961c-c3ed070cd9f9