Symfony 1.3 can't autoload the FCK rich text editor class, and Symfony 1.4 doesn't offer it at all. So this widget is now a port of the old Symfony FCK rich text editor wrapper to run as a widget instead.
I've documented a longstanding issue with FCK in widgets: FCK doesn't allow the id and name attributes
to be different, so make sure your name attribute is unique pagewide. This is usually the case
thanks to typical usages of setNameFormat() in form classes, but you can have issues if you have
multiple instances of the same form in the same page at the same time.



git-svn-id: http://svn.symfony-project.com/plugins/pkToolkitPlugin/trunk@24397 ee427ae8-e902-0410-961c-c3ed070cd9f9